+++ src/hydraulics.h
@@ -5,12 +5,12 @@
 #endif
 using namespace Rcpp;
 
-double K2Psi(double K, double Psi_extract, double ws= 3.0);
-NumericVector K2Psi(NumericVector K, NumericVector Psi_extract, double ws= 3.0);
-double Psi2K(double psi, double Psi_extract, double ws= 3.0);
-NumericVector Psi2K(double psi, NumericVector Psi_extract, double ws= 3.0);
-double averagePsi(NumericVector psi, NumericVector v, double c, double d);
-double averagePsiPool(NumericMatrix Psi, NumericMatrix RHOPcoh, double c, double d);
+double K2Psi(double K, double psi_extract, double exp_extract= 3.0);
+NumericVector K2Psi(NumericVector K, NumericVector psi_extract, double exp_extract= 3.0);
+double Psi2K(double psi, double psi_extract, double exp_extract= 3.0);
+NumericVector Psi2K(double psi, NumericVector psi_extract, double exp_extract= 3.0);
+double averagePsi(NumericVector psi, NumericVector v, double exp_extract, double psi_extract);
+double averagePsiPool(NumericMatrix Psi, NumericMatrix RHOPcoh, double exp_extract, double psi_extract);
 
 double correctConductanceForViscosity(double kxylem, double temp);